Add request statistics to Dynamite
`httpDone` is the event emitted when we've recieved all data but AWS-SDK
hasn't parsed it yet.  `complete` emitted aftere the parsing is done (or
has failed).  Buffer.byteLength is unfortunately O(n), and is about 13ms
for 6MB of data (which should be okay since requests that return 6MB of
data are expensive anyways -- the same request took ~140ms to parse).